// src/types/card.ts
export enum CardTypeCode {
  VISA = 'visa',
  MASTERCARD = 'mastercard',
  CUP = 'cup',
  DISCOVER = 'discover',
  AMEX = 'amex',
  DINERS_CLUB = 'diners_club',
  JCB = 'jcb',
  SWITCH = 'switch',
  DANKORT = 'dankort',
  VERVE = 'verve',
  RUPAY = 'rupay',
  MAESTRO = 'maestro',
  SOLO = 'solo',
  FORBRUGSFORENIGEN = 'forbrugsforeningen',
  LASER = 'laser',
  BELCARD = 'belcard',
  MIR = 'mir',
  NAPAS = 'napas',
  UZCARD = 'uzcard',
  HUMO = 'humo',
  TROY = 'troy',
  PROSTIR = 'prostir',
  KOREANLOCALCARD = 'koreanlocalcard',
  ELCART = 'elcart',
}

export interface CardType {
  code: CardTypeCode,
  regExp: RegExp,
  minLength: number,
  maxLength: number,
  luhn: boolean,
}